requirements.txt

asgiref==3.2.3

dj-database-url==0.5.0
Django==3.0
django-adaptors==0.2.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-extensions==3.0.8
django-heroku==0.3.1

gunicorn==20.0.4

Pillow==7.2.0

psycopg2==2.7.4
psycopg2-binary==2.8.6
psycopg2-pool==1.1

pylint==2.4.4
pylint-django==2.1.0
pylint-plugin-utils==0.6

pytz==2020.5

whitenoise==5.2.0
